Modularização das funções
Fiz algumas alterações nos comentários do código e modularizei as funções personalizadas, as colocando em um arquivo separado.
